Improved Keyboard usage with spacebar, backspace and shift

+++ kernel/src/math.h
@@ -2,6 +2,6 @@
 #pragma once
 
 struct Point{
-    unsigned int X;
-    unsigned int Y;
+    long X;
+    long Y;
 };
